Moved CliApp code to AppMotor
